[Embed] Finalize Embed Component
- adds an embed sub-package to the servlets package
- changes access level (protected -> public) of the TextValueDataResourceSource fields and methods
- updates access levels in classes extending TextValueDataResourceSource
