+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/internal/servlets/WorkflowModelDataSourceServlet.java
@@ -79,12 +79,12 @@ public class WorkflowModelDataSourceServlet extends SlingSafeMethodsServlet {
         }
 
         @Override
-        protected String getText() {
+        public String getText() {
             return model.getTitle();
         }
 
         @Override
-        protected String getValue() {
+        public String getValue() {
             return model.getId();
         }
     }
