+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/internal/servlets/embed/EmbeddablesDataSourceServlet.java
@@ -13,7 +13,7 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
-package com.adobe.cq.wcm.core.components.internal.servlets;
+package com.adobe.cq.wcm.core.components.internal.servlets.embed;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -38,6 +38,7 @@ import org.apache.sling.jcr.resource.api.JcrResourceConstants;
 import org.jetbrains.annotations.NotNull;
 import org.osgi.service.component.annotations.Component;
 
+import com.adobe.cq.wcm.core.components.internal.servlets.TextValueDataResourceSource;
 import com.adobe.cq.wcm.core.components.models.Embed;
 import com.adobe.granite.ui.components.ds.DataSource;
 import com.adobe.granite.ui.components.ds.SimpleDataSource;
@@ -124,12 +125,12 @@ public class EmbeddablesDataSourceServlet extends SlingSafeMethodsServlet {
         }
 
         @Override
-        protected String getText() {
+        public String getText() {
             return description.getTitle();
         }
 
         @Override
-        protected String getValue() {
+        public String getValue() {
             return description.getResourceType();
         }
     }
